+++ app/src/main/java/woowacourse/shopping/model/data/DummyProducts.kt
@@ -1,4 +1,6 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.model.data
+
+import woowacourse.shopping.model.Product
 
 val MAC_BOOK =
     Product(
