+++ app/src/main/java/woowacourse/shopping/model/data/ProductsImpl.kt
@@ -1,4 +1,6 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.model.data
+
+import woowacourse.shopping.model.Product
 
 object ProductsImpl : ProductDao {
     private const val EXCEPTION_INVALID_ID = "Movie not found with id: %d"
