+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModelFactory.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.ui.cart
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.model.CartDao
+import woowacourse.shopping.model.data.CartDao
 
 class CartViewModelFactory(
     private val cartDao: CartDao,
