+++ app/src/test/java/woowacourse/shopping/model/data/ProductsImplTest.kt
@@ -1,8 +1,9 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.model.data
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import woowacourse.shopping.model.Product
 
 class ProductsImplTest {
     @BeforeEach
