+++ app/src/test/java/woowacourse/shopping/ui/cart/CartViewModelTest.kt
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import woowacourse.shopping.InstantTaskExecutorExtension
 import woowacourse.shopping.getOrAwaitValue
-import woowacourse.shopping.model.CartsImpl
 import woowacourse.shopping.model.Product
+import woowacourse.shopping.model.data.CartsImpl
 
 @ExtendWith(InstantTaskExecutorExtension::class)
 class CartViewModelTest {
