+++ app/src/test/java/woowacourse/shopping/ui/products/ProductContentsViewModelTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import woowacourse.shopping.InstantTaskExecutorExtension
 import woowacourse.shopping.getOrAwaitValue
-import woowacourse.shopping.model.ProductsImpl
+import woowacourse.shopping.model.data.ProductsImpl
 
 @ExtendWith(InstantTaskExecutorExtension::class)
 class ProductContentsViewModelTest {
